Factor out all the duplication from these tests
Still needs improvement
